8299528: IR test: TestEor3AArch64.java fails on aarch64
The IR test - TestEor3AArch64.java fails during IR verification as the
IR nodes to be matched have not been updated to the newer IR framework
introduced in the JDK source with this PR -
https://github.com/openjdk/jdk/pull/10695.

This patch updates the relevant IR nodes in the tests accordingly.

The test passes successfully on -
1. an aarch64 machine with vector size of 128-bit that supports the Neon
eor3 instruction
2. an aarch64 simulator with vector size of 256-bit that supports the SVE2
feature which includes the SVE2 eor3 instruction.
Generation of IR Nodes - XOR3_NEON and XOR3_SVE have thus been verified
on the above mentioned machines.

This is a trivial patch related to a test and has minimal impact with
low risk and maybe considered to be merged with JDK20 in RDP1.